#27 refactor: declare global GameObjectFactory types
