+++ frontend/src/PhaserGame.ts
@@ -8,9 +8,8 @@ const config: Phaser.Types.Core.GameConfig = {
   backgroundColor: '#64748b',
   pixelArt: true,
   scale: {
-    mode: Phaser.Scale.ScaleModes.FIT,
-    width: 800,
-    height: 600,
+    width: 700,
+    height: 700,
     autoCenter: Phaser.Scale.CENTER_BOTH,
   },
   physics: {
