+++ frontend/src/types/GameObjectFactory.d.ts
@@ -1,0 +1,16 @@
+export {};
+
+// declare myPlayer type in GameObjectFactory
+declare global {
+  namespace Phaser.GameObjects {
+    interface GameObjectFactory {
+      myPlayer: (
+        x: number,
+        y: number,
+        texture: string,
+        frame?: string | number,
+        options?: Phaser.Types.Physics.Matter.MatterBodyConfig
+      ) => MyPlayer;
+    }
+  }
+}
